+++ Plumber.build.ps1
@@ -5,21 +5,20 @@ if (-not $module) {
     $module = Import-Module (Join-Path $PSScriptRoot 'Plumber.psd1') -Force -PassThru
 }
 
+Import-Module Plumber.Release -RequiredVersion 0.1.0 -Force
+
 . (Get-PlumberTaskLoader) -Config @{
     ModuleManifest = 'Plumber.psd1'
     ExcludePaths = @{
-        Backticks        = @('out/**')
-        JSON             = @('out/**')
-        JSONSchema       = @('out/**')
-        LineLength       = @('out/**')
-        PSScriptAnalyzer = @('Tests/Assets/TaskHelp/InvalidPowerShell.ps1', 'out/**')
-        ToDo             = @('out/**')
-        YAML             = @('out/**')
+        PSScriptAnalyzer = @('Tests/Assets/TaskHelp/InvalidPowerShell.ps1')
     }
     LocalTasks = @(
         'LocalTasks/ValidateTaskHelp.ps1'
     )
 }
 
-. (Join-Path $PSScriptRoot 'build/Publish.ps1') -ModuleBuildExtraItems @('Tasks', 'docs')
+. (Get-PlumberReleaseTaskLoader) -Config @{
+    ModuleManifest = 'Plumber.psd1'
+}
+
 . (Join-Path $PSScriptRoot 'LocalTasks/GenerateDocs.ps1')